+++ common/src/main/java/net/mehvahdjukaar/supplementaries/common/block/cannon/BoatReferenceFrame.java
@@ -30,7 +30,7 @@ public class BoatReferenceFrame extends EntityReferenceFrame{
 
     @Override
     public boolean shouldRotatePlayerFaceWhenManeuvering() {
-        return true;
+        return false;
     }
 
     @Override