+++ common/src/main/java/net/mehvahdjukaar/supplementaries/common/entities/goals/PlundererAICommon.java
@@ -101,7 +101,7 @@ public final class PlundererAICommon {
 
             //sync
             tile.setChanged();
-            tile.syncToClients();
+            tile.syncToClients(false);
         }
     }