+++ common/src/main/java/net/mehvahdjukaar/supplementaries/common/network/SyncCannonPacket.java
@@ -15,7 +15,7 @@ import net.minecraft.world.level.block.entity.BlockEntity;
 import org.joml.Quaternionf;
 
 public record SyncCannonPacket(
-        Quaternionf rotation, byte firePower, boolean fire, boolean stopControlling,
+        Quaternionf rotation, byte firePower, boolean ignite, boolean stopControlling,
         TileOrEntityTarget target) implements Message {
 
     public static final TypeAndCodec<RegistryFriendlyByteBuf, SyncCannonPacket> CODEC = Message.makeType(
@@ -30,7 +30,7 @@ public record SyncCannonPacket(
     public void write(RegistryFriendlyByteBuf buf) {
         ByteBufCodecs.QUATERNIONF.encode(buf, rotation);
         buf.writeByte(this.firePower);
-        buf.writeBoolean(this.fire);
+        buf.writeBoolean(this.ignite);
         buf.writeBoolean(this.stopControlling);
         this.target.write(buf);
     }
@@ -44,12 +44,12 @@ public record SyncCannonPacket(
         BlockEntity be = this.target.findTileOrContainedTile(level);
         if (be instanceof CannonBlockTile cannon) {
             if (cannon.canBeUsedBy(BlockPos.containing(cannon.getGlobalPosition(1)), player)) {
-                cannon.setAttributes(this.rotation, this.firePower, this.fire, player);
+                cannon.setAttributes(this.rotation, this.firePower, this.ignite, player);
                 cannon.setChanged();
                 if (stopControlling) {
                     cannon.setCurrentUser(null);
                 }
-                if (!level.isClientSide) cannon.syncToClients();
+                if (!level.isClientSide) cannon.syncToClients(ignite);
             } else {
                 Supplementaries.LOGGER.warn("Player tried to control cannon {} without permission: {}", player.getName().getString(), this.target);
             }